I added some keywords
I added some effects in the bullet script that will need to updated on later. Most of the code will likely be reliant on code in the enemy script if someone can work on that later. The reason why I am leaving it here is because I need to work of cards and the deck.